pyproject.toml
[project]
name = "skill-aware-router"
version = "0.1.0"
description = "A Skills.sh Orchestration Layer using FastMCP 3.0 — routes tasks to the best model+skill pair."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=3.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "google-genai>=1.0",
    "anthropic>=0.40",
    "watchfiles>=1.0",
]

[project.scripts]
skill-router = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
